Consolidate gm corner and tutorial view components
