+++ services/app/src/components/views/GMCornerContainer.tsx
@@ -7,12 +7,14 @@ import {getContentSets} from 'app/actions/Settings';
 import {GM_QUESTS} from 'app/Constants';
 import {AppState} from 'app/reducers/StateTypes';
 import {Quest} from 'shared/schema/Quests';
-import GMCorner, {DispatchProps, StateProps} from './GMCorner';
+import QuestListCard, {DispatchProps, StateProps} from '../base/QuestListCard';
 
 const mapStateToProps = (state: AppState): StateProps => {
   return {
     quests: GM_QUESTS,
     contentSets: getContentSets(state.settings, state.multiplayer),
+    title: 'Tools',
+    icon: 'gm_corner',
   };
 };
 
@@ -31,6 +33,6 @@ const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
 const GMCornerContainer = connect(
   mapStateToProps,
   mapDispatchToProps
-)(GMCorner);
+)(QuestListCard);
 
 export default GMCornerContainer;
