+++ services/app/src/components/views/TutorialsContainer.tsx
@@ -6,13 +6,14 @@ import {previewQuest} from '../../actions/Quest';
 import {getContentSets} from '../../actions/Settings';
 import {TUTORIAL_QUESTS} from '../../Constants';
 import {AppState} from '../../reducers/StateTypes';
-import Tutorials, {DispatchProps, StateProps} from './Tutorials';
+import QuestListCard, {DispatchProps, StateProps} from '../base/QuestListCard';
 
 const mapStateToProps = (state: AppState): StateProps => {
   return {
     quests: TUTORIAL_QUESTS,
-    settings: state.settings,
     contentSets: getContentSets(state.settings, state.multiplayer),
+    title: 'tutorials',
+    icon: 'helper',
   };
 };
 
@@ -30,6 +31,6 @@ const mapDispatchToProps = (dispatch: Redux.Dispatch<any>): DispatchProps => {
 const TutorialsContainer = connect(
   mapStateToProps,
   mapDispatchToProps
-)(Tutorials);
+)(QuestListCard);
 
 export default TutorialsContainer;
